+++ tests/AnomalyDetectors/OneClassSVMTest.php
@@ -1,23 +1,24 @@
 <?php
 
-namespace Rubix\ML\Tests\Regressors;
+namespace Rubix\ML\Tests\AnomalyDetectors;
 
 use Rubix\ML\Learner;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Unlabeled;
-use Rubix\ML\Regressors\KDNRegressor;
-use Rubix\ML\Kernels\Distance\Minkowski;
-use Rubix\ML\Datasets\Generators\HalfMoon;
+use Rubix\ML\Kernels\SVM\Polynomial;
+use Rubix\ML\Datasets\Generators\Blob;
+use Rubix\ML\Datasets\Generators\Circle;
+use Rubix\ML\AnomalyDetectors\OneClassSVM;
+use Rubix\ML\Datasets\Generators\Agglomerate;
 use PHPUnit\Framework\TestCase;
 use InvalidArgumentException;
 use RuntimeException;
 
-class KDNRegressorTest extends TestCase
+class OneClassSVMTest extends TestCase
 {
     const TRAIN_SIZE = 300;
-    const TEST_SIZE = 5;
-    const TOLERANCE = 10;
+    const TEST_SIZE = 10;
 
     protected $generator;
 
@@ -25,14 +26,17 @@ class KDNRegressorTest extends TestCase
 
     public function setUp()
     {
-        $this->generator = new HalfMoon(4., -7., 1., 90, 0.02);
+        $this->generator = new Agglomerate([
+            0 => new Blob([0., 0.], 0.5),
+            1 => new Circle(0., 0., 8., 0.1),
+        ], [0.9, 0.1]);
 
-        $this->estimator = new KDNRegressor(3, 10, new Minkowski(3.0));
+        $this->estimator = new OneClassSVM(0.1, new Polynomial(4, 0.01), true, 1e-3);
     }
 
-    public function test_build_regressor()
+    public function test_build_detector()
     {
-        $this->assertInstanceOf(KDNRegressor::class, $this->estimator);
+        $this->assertInstanceOf(OneClassSVM::class, $this->estimator);
         $this->assertInstanceOf(Learner::class, $this->estimator);
         $this->assertInstanceOf(Persistable::class, $this->estimator);
         $this->assertInstanceOf(Estimator::class, $this->estimator);
@@ -40,25 +44,18 @@ class KDNRegressorTest extends TestCase
 
     public function test_estimator_type()
     {
-        $this->assertEquals(Estimator::REGRESSOR, $this->estimator->type());
+        $this->assertEquals(Estimator::DETECTOR, $this->estimator->type());
     }
 
     public function test_train_predict()
     {
-        $this->estimator->train($this->generator->generate(self::TRAIN_SIZE));
-
         $testing = $this->generator->generate(self::TEST_SIZE);
 
+        $this->estimator->train($this->generator->generate(self::TRAIN_SIZE));
+
         foreach ($this->estimator->predict($testing) as $i => $prediction) {
-            $this->assertEquals($testing->label($i), $prediction, '', self::TOLERANCE);
+            $this->assertEquals($testing->label($i), $prediction);
         }
-    }
-
-    public function test_train_with_unlabeled()
-    {
-        $this->expectException(InvalidArgumentException::class);
-
-        $this->estimator->train(Unlabeled::quick());
     }
 
     public function test_predict_untrained()